Split REST connectors and move to own package
