+++ qiskit/providers/ibmq/api/rest/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2019, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+"""REST adaptors for the IBM Q Api version 2."""
+
+from .auth import Auth
+from .root import Api
